Vader_Secret_Keeper1/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic>=2
